make ValidatedRequest compatible with MockExporter + remove ApodiniRequest entirely
