+++ Sources/Apodini/Semantic Model Builder/RequestInjectable.swift
@@ -31,12 +31,12 @@ extension RequestInjectableVisitor {
 }
 
 extension Handler {
-    func extractRequestInjectables() -> [String: RequestInjectable] {
+    func extractRequestInjectables() -> [(String, RequestInjectable)] {
         Apodini.extractRequestInjectables(from: self)
     }
 }
 extension AnyResponseTransformer {
-    func extractRequestInjectables() -> [String: RequestInjectable] {
+    func extractRequestInjectables() -> [(String, RequestInjectable)] {
         Apodini.extractRequestInjectables(from: self)
     }
 }
